requirements.txt
os-ken==2.7.0
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.4
python-socketio==5.10.0
eventlet==0.35.2
requests==2.31.0
nltk==3.8.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
textblob==0.17.1
groq==0.9.0
python-dotenv==1.0.0
# Commented: netaddr, pyyaml, webob, spacy, transformers, torch, scikit-learn, redis
